check if the file exists before unlinking

+++ src/imagico.js
@@ -16,7 +16,9 @@ function ImagicoElevationDownloader(cacheDir, options) {
 ImagicoElevationDownloader.prototype.download = function(tileKey, latLng, cb) {
     var cleanup = function() {
             delete this._downloads[tileKey];
-            fs.unlinkSync(tempPath);
+            if (fs.existsSync(tempPath)) {
+                fs.unlinkSync(tempPath);
+            }
         }.bind(this),
         download = this._downloads[tileKey],
         tempPath,
